[add]: LayoutType enum 파일 분리
